requirements.txt
viberbot~=1.0.12
flask~=3.0.0
SQLAlchemy~=2.0.23
python-dotenv~=1.0.0
pymysql~=1.1.0